Update e2e tests to include essential plugins
