+++ pkg/constants/defaults.go
@@ -20,7 +20,6 @@ const (
 	// from where the CLI will discover the plugins
 	TanzuCLIDefaultCentralPluginDiscoveryImage = "projects.registry.vmware.com/tanzu_cli/plugins/plugin-inventory:latest"
 
-	// DefaultCLIEssentialsPluginGroupName  name of the default essentials plugin group which is used to install essential plugins
-	// TODO: Update the group name
-	DefaultCLIEssentialsPluginGroupName = "vmware-tzcli/essentials"
+	// DefaultCLIEssentialsPluginGroupName  name of the essentials plugin group which is used to install essential plugins
+	DefaultCLIEssentialsPluginGroupName = "vmware-tanzucli/essentials"
 )
