+++ test/e2e/airgapped/plugins.go
@@ -9,8 +9,6 @@ import "github.com/vmware-tanzu/tanzu-cli/test/e2e/framework"
 // TODO(anujc): Remove the hardcoding of the plugins within the plugin-groups here once the `tanzu plugin group get`
 // command is implemented. We can use the output of `tanzu plugin group get` command with the original repository
 // and use the created plugin map to do the validation of plugins availability after plugin gets migrated to new repository.
-//
-//nolint:dupl
 var pluginsForPGTKG001 = []*framework.PluginInfo{
 	{Name: "cluster", Target: "kubernetes", Version: "v0.0.1", Description: "Desc for cluster"},
 	{Name: "feature", Target: "kubernetes", Version: "v0.0.1", Description: "Desc for feature"},
@@ -18,10 +16,8 @@ var pluginsForPGTKG001 = []*framework.PluginInfo{
 	{Name: "management-cluster", Target: "kubernetes", Version: "v0.0.1", Description: "Desc for management-cluster"},
 	{Name: "package", Target: "kubernetes", Version: "v0.0.1", Description: "Desc for package"},
 	{Name: "secret", Target: "kubernetes", Version: "v0.0.1", Description: "Desc for secret"},
-	{Name: "telemetry", Target: "kubernetes", Version: "v0.0.1", Description: "Desc for telemetry"},
 }
 
-//nolint:dupl
 var pluginsForPGTKG999 = []*framework.PluginInfo{
 	{Name: "cluster", Target: "kubernetes", Version: "v9.9.9", Description: "Desc for cluster"},
 	{Name: "feature", Target: "kubernetes", Version: "v9.9.9", Description: "Desc for feature"},
@@ -29,7 +25,6 @@ var pluginsForPGTKG999 = []*framework.PluginInfo{
 	{Name: "management-cluster", Target: "kubernetes", Version: "v9.9.9", Description: "Desc for management-cluster"},
 	{Name: "package", Target: "kubernetes", Version: "v9.9.9", Description: "Desc for package"},
 	{Name: "secret", Target: "kubernetes", Version: "v9.9.9", Description: "Desc for secret"},
-	{Name: "telemetry", Target: "kubernetes", Version: "v9.9.9", Description: "Desc for telemetry"},
 }
 
 //nolint:dupl
@@ -79,4 +74,5 @@ var pluginsForPGTMC999 = []*framework.PluginInfo{
 var pluginsNotInAnyPG999 = []*framework.PluginInfo{
 	{Name: "isolated-cluster", Target: "global", Version: "v9.9.9", Description: "Desc for isolated-cluster"},
 	{Name: "pinniped-auth", Target: "global", Version: "v9.9.9", Description: "Desc for pinniped-auth"},
+	{Name: "telemetry", Target: "global", Version: "v9.9.9", Description: "Desc for telemetry"},
 }
