+++ test/e2e/framework/framework.go
@@ -33,6 +33,9 @@ var PluginsForLifeCycleTests []*PluginInfo
 // PluginGroupsForLifeCycleTests is list of plugin groups (which are published in local central repo) used in plugin group life cycle test cases
 var PluginGroupsForLifeCycleTests []*PluginGroup
 
+// EssentialPluginGroups is list of essential plugin groups
+var EssentialPluginGroups []*PluginGroup
+
 // CLICoreDescribe annotates the test with the CLICore label.
 func CLICoreDescribe(text string, body func()) bool {
 	return ginkgo.Describe(CliCore+text, body)
@@ -110,8 +113,18 @@ func init() {
 	_ = CreateDir(FullPathForTempDir)
 
 	// TODO:cpamuluri: need to move plugins info to configuration file with positive and negative use cases - github issue: https://github.com/vmware-tanzu/tanzu-cli/issues/122
-	PluginsForLifeCycleTests = []*PluginInfo{{Name: "cluster", Target: "kubernetes", Version: "v9.9.9", Description: "cluster functionality"}, {Name: "cluster", Target: "mission-control", Version: "v9.9.9", Description: "cluster functionality"}, {Name: "pinniped-auth", Target: "global", Version: "v9.9.9", Description: "pinniped-auth functionality"}}
+	PluginsForLifeCycleTests = []*PluginInfo{
+		{Name: "cluster", Target: "kubernetes", Version: "v9.9.9", Description: "cluster functionality"},
+		{Name: "cluster", Target: "mission-control", Version: "v9.9.9", Description: "cluster functionality"},
+		{Name: "pinniped-auth", Target: "global", Version: "v9.9.9", Description: "pinniped-auth functionality"},
+		{Name: "telemetry", Target: "global", Version: "v9.9.9", Description: "telemetry functionality"},
+	}
 
 	// TODO:cpamuluri: need to move Plugin Groups to configuration file with positive and negative use cases - github issue: https://github.com/vmware-tanzu/tanzu-cli/issues/122
-	PluginGroupsForLifeCycleTests = []*PluginGroup{{Group: "vmware-tkg/default", Latest: "v9.9.9", Description: "Desc for vmware-tkg/default:v9.9.9"}, {Group: "vmware-tmc/tmc-user", Latest: "v9.9.9", Description: "Desc for vmware-tmc/tmc-user:v9.9.9"}}
+	PluginGroupsForLifeCycleTests = []*PluginGroup{
+		{Group: "vmware-tkg/default", Latest: "v9.9.9", Description: "Desc for vmware-tkg/default:v9.9.9"},
+		{Group: "vmware-tmc/tmc-user", Latest: "v9.9.9", Description: "Desc for vmware-tmc/tmc-user:v9.9.9"},
+		{Group: "vmware-tzcli/essentials", Latest: "v9.9.9", Description: "Desc for vmware-tzcli/essentials:v9.9.9"},
+	}
+	EssentialPluginGroups = []*PluginGroup{{Group: "vmware-tzcli/essentials", Latest: "v9.9.9", Description: "Desc for vmware-tzcli/essentials:v9.9.9"}}
 }
